+++ src/spec/decorators/testDecorator/multipleTestsTestDecoratorTestSuite.ts
@@ -1,23 +1,15 @@
 import { test } from '../../../testyCore';
+import { testSuite } from '../../../lib/decorators/testSuite.decorator';
 
+@testSuite('Multiple Test Test Decorator Test Suite')
 export class MultipleTestTestDecoratorTestSuite {
 
-    public numberOfRunsTest1: number = 0;
-    public numberOfRunsTest2: number = 0;
-    public numberOfRunsTest3: number = 0;
+    @test('My first test')
+    private test1() { }
 
-    @test('My test 1.')
-    private test1() {
-        ++this.numberOfRunsTest1;
-    }
+    @test('My second test')
+    private test2() { }
 
-    @test('My test 2.')
-    private test2() {
-        ++this.numberOfRunsTest2;
-    }
-
-    @test('My test 3.')
-    private test3() {
-        ++this.numberOfRunsTest3;
-    }
+    @test('My third test')
+    private test3() { }
 }