+++ src/spec/decorators/testSuiteTestsBase.ts
@@ -1,0 +1,18 @@
+import { beforeEach } from '../../lib/decorators/beforeEach.decorator';
+import { Report } from '../../lib/reporting/report/report';
+import { TestSuite } from '../../lib/tests/testSuite';
+import { TestVisitor } from '../../lib/tests/visitors/testVisitor';
+import { TestRunnerVisitor } from '../../lib/tests/visitors/testRunnerVisitor';
+
+export class TestSuiteTestsBase {
+    protected visitor: TestVisitor<Report>;
+
+    @beforeEach()
+    private beforeEach() {
+        this.visitor = new TestRunnerVisitor();
+    }
+
+    protected getTestSuiteInstance(testClass: any): TestSuite {
+        return testClass.__testSuiteInstance;
+    }
+}