+++ src/spec/tests/visitors/timeoutTestDecoratorTestSuite.ts
@@ -1,5 +1,7 @@
 import { test } from '../../../testyCore';
+import { testSuite } from '../../../lib/decorators/testSuite.decorator';
 
+@testSuite('Timeout Test Decorator Test Suite')
 export class TimeoutTestDecoratorTestSuite {
     @test('My test with test cases', undefined, 10)
     private async tests(n: number) {